#77 Minor bug fixes; had an error in routes, undef when making a new team, added translations for Datatable (more later)

+++ app/Models/Team.php
@@ -252,7 +252,9 @@ class Team extends IconFileModel
             /** @var $item Team */
 
             // Delete icons
-            $item->iconfile->delete();
+            if ($item->iconfile !== null) {
+                $item->iconfile->delete();
+            }
 
             // Remove all users associated with this team
             TeamUser::where('team_id', $item->id)->delete();
